+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/SchemaBuilders.java
@@ -216,8 +216,8 @@ public final class SchemaBuilders {
      * Returns a new {@link SnowballAnalyzerBuilder} for the specified language and stopwords.
      *
      * @param language The language. The supported languages are English, French, Spanish, Portuguese, Italian,
-     * Romanian, German, Dutch, Swedish, Norwegian, Danish, Russian, Finnish, Hungarian and Turkish.
-     * Basque and Catalan.
+     * Romanian, German, Dutch, Swedish, Norwegian, Danish, Russian, Finnish, Hungarian and Turkish. Basque and
+     * Catalan.
      * @param stopwords the comma separated stopwords {@code String} list
      * @return a new snowball analyzer builder
      */